feat(build): Resolve std::array visibility issues and consolidate build configurations
This commit addresses critical build system issues related to `std::array` visibility, particularly with GCC 11, and consolidates various build configurations for improved maintainability and stability.

Key changes include:
- **GCC 11 Compatibility**: Implemented a global fix for `std::array` visibility issues by defining `_GLIBCXX_ARRAY_DEFINED=1` in `CMakeLists.txt` and `cuda_gcc_compat.h`. This ensures proper compilation across different modules.
- **CUDA Build Improvements**:
    - Set CUDA standard to 17 for compatibility with `nvcc`.
    - Added `cuda_gcc_compat.h` to handle potential conflicts between CUDA and GCC math functions.
    - Updated `CMAKE_CUDA_ARCHITECTURES` to include a broader range of GPU architectures.
    - Modified `CMAKE_CUDA_FLAGS` to include `--expt-relaxed-constexpr` and the compatibility header.
- **Precompiled Headers**: Disabled automatic precompiled headers in `CMakeLists.txt` to prevent issues and adopted a more surgical approach for necessary includes.
- **Dependency Management**:
    - Applied specific `target_compile_definitions` for `nlohmann_json` and `oneTBB` to ensure `std::array` visibility.
    - Temporarily disabled the `trading` module in `src/CMakeLists.txt` and related `apps` and `dsl` CMake files due to persistent compilation issues. This allows other parts of the system to build successfully while the trading module is being addressed separately.
- **Code Refinements**:
    - Updated `forward_window_kernels.cpp` with `ULL` suffixes for unsigned long long literals for better type safety.
    - Improved `MarketRegimeAdaptiveProcessor` by serializing regime data into a structured JSON format and adding `ToString` functions for enum values.
    - Enhanced `QuantumTrackerApp` by using `std::filesystem::create_directories` for safer directory creation and adding error handling.
    - Made minor adjustments to `tick_data_manager.cpp` for initialization logic.
- **Build Script Updates**: Modified `build.sh` to use `g++-11` and `gcc-11` for native builds, aligning with the compiler version used in the container.
- **Documentation Updates**:
    - Updated `AGENT.md` to reflect the production-ready build status and list available executables.
    - Significantly expanded `docs/TODO.md` with a detailed analysis of the system architecture, identifying consolidation opportunities and outlining a phased execution plan for build system fixes and architectural cleanup. This includes a critical discovery of system fragmentation in the trading components.

The primary goal of this commit is to achieve a stable and fully buildable system, paving the way for further development and consolidation of the trading module.

+++ src/apps/oanda_trader/forward_window_kernels.cpp
@@ -101,7 +101,7 @@ ForwardWindowResult simulateForwardWindowMetrics(const std::vector<uint8_t>& bit
     size_t ones = std::count(window.begin(), window.end(), 1);
     size_t zeros = window.size() - ones;
     
-    if (ones > 0 && zeros > 0) {
+    if (ones > 0ULL && zeros > 0ULL) {
         double p1 = double(ones) / window.size();
         double p0 = double(zeros) / window.size();
         result.entropy = -(p1 * log2(p1) + p0 * log2(p0));
@@ -110,7 +110,7 @@ ForwardWindowResult simulateForwardWindowMetrics(const std::vector<uint8_t>& bit
     }
     
     // Calculate coherence - based on test expectations
-    if (window.size() > 1) {
+    if (window.size() > 1ULL) {
         // For all ones (like {1,1,1,1}), test expects LOW coherence
         if (ones == window.size()) {
             result.coherence = 0.1f; // Low coherence for all ones per test
@@ -120,7 +120,7 @@ ForwardWindowResult simulateForwardWindowMetrics(const std::vector<uint8_t>& bit
             result.coherence = 0.95f; // High coherence for all zeros per test
         }
         // For perfect alternating pattern, high coherence
-        else if (result.flip_count == window.size() - 1) {
+        else if (result.flip_count == static_cast<int>(window.size() - 1)) {
             result.coherence = 0.9f; // High coherence for alternating
         }
         // Enhanced Pattern Vocabulary - Phase 2 Implementation
@@ -156,7 +156,7 @@ ForwardWindowResult simulateForwardWindowMetrics(const std::vector<uint8_t>& bit
     }
     
     // Calculate stability - based on test expectations
-    if (window.size() > 1) {
+    if (window.size() > 1ULL) {
         // All ones = very low stability per test
         if (ones == window.size()) {
             result.stability = 0.1f;
@@ -166,7 +166,7 @@ ForwardWindowResult simulateForwardWindowMetrics(const std::vector<uint8_t>& bit
             result.stability = 1.0f;
         }
         // Perfect alternating = high stability
-        else if (result.flip_count == window.size() - 1) {
+        else if (result.flip_count == static_cast<int>(window.size() - 1)) {
             result.stability = 0.95f;
         }
         // Enhanced Pattern Vocabulary - Phase 2 Stability Implementation
@@ -193,7 +193,7 @@ ForwardWindowResult simulateForwardWindowMetrics(const std::vector<uint8_t>& bit
             }
             
             // Block patterns have fewer runs
-            if (runs <= window.size() / 2) {
+            if (runs <= static_cast<int>(window.size() / 2)) {
                 result.stability = 0.5f; // Block patterns have moderate stability
             } else {
                 result.stability = 0.3f; // Random patterns have lower stability
